Chromium should call restrictScaleFactorToInitialScaleIfNotUserScalable unless/until userScalable is supported directly. https://bugs.webkit.org/show_bug.cgi?id=91110
Reviewed by Tony Chang.

Source/WebKit/chromium:

Prior to this patch, we were ignoring the user-scalable property of
viewport meta tags. This patch enforces the user-scalable property by
claming the minimum-scale and maximum-scale values, effectively
preventing user scaling. This approach isn't perfect, as discussed in
https://bugs.webkit.org/show_bug.cgi?id=70609, but it is an improvement
over the status quo.

This approach is also used by EFL, GTK, and Qt.

* src/ChromeClientImpl.cpp:
(WebKit::ChromeClientImpl::dispatchViewportPropertiesDidChange):

LayoutTests:

Test that the scale limits are clamped when user-scalable is set to no.

* fast/viewport/viewport-limits-adjusted-for-no-user-scale-control-expected.txt: Added.
* fast/viewport/viewport-limits-adjusted-for-no-user-scale-control.html: Added.
* fast/viewport/viewport-limits-adjusted-for-no-user-scale-expected.txt: Added.
* fast/viewport/viewport-limits-adjusted-for-no-user-scale.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@127704 268f45cc-cd09-0410-ab3c-d52691b4dbfc