Web Inspector: Persistent handle referenced from ScriptWrappable is double counted https://bugs.webkit.org/show_bug.cgi?id=96483
Reviewed by Alexander Pavlov.

Source/WebCore:

* bindings/v8/ScriptWrappable.h:
(WebCore::ScriptWrappable::reportMemoryUsage): the handle is a part of an
array where all such handles are allocated and should not be counted here
second time. In order to make the clang plugin that validate memory instrumentation
happy we report it here as weak pointer (no-op).

Source/WTF:

* wtf/MemoryInstrumentation.h:
(WebCore::MemoryClassInfo::addWeakPointer): this method is expected to be
used on fields that are pointers to objects which are parts of bigger memory
blocks (field of another object, element in an array, object allocated in a
memory arena etc.). We don't want to count such objects' memory separately
from their owners but in order to be able to validates the memory instrumentation
with clang plugin we need to make sure all fields in instrumented objects
are reported.
(MemoryClassInfo):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@128283 268f45cc-cd09-0410-ab3c-d52691b4dbfc